# code/rhserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/index/insertIndex": {
            "post": {
                "description": "insert stock index information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "insertIndex",
                "parameters": [
                    {
                        "description": "index information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.IndexInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.IndexInfo"
                        }
                    }
                }
            }
        },
        "/v1/index/insertIndexHq": {
            "post": {
                "description": "insert stock index hq data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "insertIndexHq",
                "parameters": [
                    {
                        "description": "index information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.IndexHq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.IndexHq"
                        }
                    }
                }
            }
        },
...
        "/v1/index/queryIndexCoStockByCond": {
            "get": {
                "description": "查询指数成分股",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "queryIndexCoStockByCond",
                "parameters": [
                    {
                        "type": "string",
                        "description": "index code",
                        "name": "index_code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "publish date",
                        "name": "public_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.IndexStock"
                        }
                    }
                }
            }
        },
        "/v1/index/queryIndexHqByCode": {
            "get": {
                "description": "查询指数行情数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "queryIndexHqByCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "index code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.IndexHq"
                        }
                    }
                }
            }
        },
        "/v1/index/queryIndexInfoByCode": {
            "get": {
                "description": "查询股票指数基本信息，行情有哪些指数，指数的代码，名称，更新时间等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "queryIndexInfoByCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "index code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.IndexInfo"
                        }
                    }
                }
            }
        },
        "/v1/index/updateIndex": {
            "post": {
                "description": "update stock index information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "updateIndex",
                "parameters": [
                    {
                        "description": "index information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.IndexInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.IndexInfo"
                        }
                    }
                }
            }
        },
        "/v1/index/updateIndexHq": {
            "post": {
                "description": "update stock index hq data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock index"
                ],
                "summary": "updateIndexHq",
                "parameters": [
                    {
                        "description": "index hq",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.IndexHq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.IndexHq"
                        }
                    }
                }
            }
        },
        "/v1/stock/insertRealTimeHq": {
            "post": {
                "description": "insert stock real-time data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "insertRealTimeHq",
                "parameters": [
                    {
                        "description": "stock real-time data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                }
            }
        },
        "/v1/stock/insertStockHisHq": {
            "post": {
                "description": "insert stock history data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "insertStockHisHq",
                "parameters": [
                    {
                        "description": "stock history data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                }
            }
        },
        "/v1/stock/insertStockInfo": {
            "post": {
                "description": "insert stock information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "insertStockInfo",
                "parameters": [
                    {
                        "description": "stock information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockInfo"
                        }
                    }
                }
            }
        },
...
        "/v1/stock/queryRealTimeHqByCode": {
            "get": {
                "description": "查询股票实时行情数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "queryRealTimeHqByCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stock code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query start time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query end time",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                }
            }
        },
        "/v1/stock/queryStockHisHq": {
            "get": {
                "description": "查询股票实时行情数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "queryStockHisHq",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stock code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query start date",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query end date",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query period, day, week, month",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query adjust, qfq, hfq, empty for no adjust",
                        "name": "adjust",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "query order, order by createdate, 1 for asc, 0 for desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                }
            }
        },
        "/v1/stock/queryStockInfoByCode": {
            "get": {
                "description": "根据code查询股票基本信息，股票基本信息包含股票代码，名称，市值，流通，所属行业等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "queryStockInfoByCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stock code",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockInfo"
                        }
                    }
                }
            }
        },
        "/v1/stock/updateRealTimeHq": {
            "post": {
                "description": "update stock real-time data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "updateRealTimeHq",
                "parameters": [
                    {
                        "description": "stock real-time data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                }
            }
        },
        "/v1/stock/updateStockHisHq": {
            "post": {
                "description": "update stock history data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "updateStockHisHq",
                "parameters": [
                    {
                        "description": "stock history data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                }
            }
        },
        "/v1/stock/updateStockInfo": {
            "post": {
                "description": "update stock information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "updateStockInfo",
                "parameters": [
                    {
                        "description": "stock information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "mongo.IndexHq": {
            "type": "object",
            "required": [
                "amount",
                "chgamount",
                "close",
                "code",
                "high",
                "low",
                "name",
                "open",
                "perchg",
                "volume"
            ],
            "properties": {
                "adjust": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "chgamount": {
                    "type": "number"
                },
                "close": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "low": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "open": {
                    "type": "number"
                },
                "perchg": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
        "mysql.IndexInfo": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "publish_date": {
                    "type": "string"
                },
                "update_date": {
                    "type": "string"
                }
            }
        },
        "mysql.IndexStock": {
            "type": "object",
            "required": [
                "code",
                "index_code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "in_date": {
                    "type": "string"
                },
                "index_code": {
                    "description": "代码",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "update_date": {
                    "type": "string"
                }
            }
        },
        "mysql.StockHisHq": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "adjust": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "chgamount": {
                    "type": "number"
                },
                "close": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "low": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "open": {
                    "type": "number"
                },
                "pchg": {
                    "type": "number"
                },
                "period": {
                    "type": "string"
                },
                "turnover": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
        "mysql.StockInfo": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "circul_cap": {
                    "type": "number"
                },
                "circul_shares": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "his_update_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "industry": {
                    "type": "string"
                },
                "market_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "total_cap": {
                    "type": "number"
                },
                "total_shares": {
                    "type": "number"
                }
            }
        },
        "mysql.StockRealTimeData": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "cap": {
                    "type": "number"
                },
                "chgamount": {
                    "type": "number"
                },
                "chginyear": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "createtime": {
                    "type": "string"
                },
                "fcap": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "low": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "open": {
                    "type": "number"
                },
                "pchg": {
                    "type": "number"
                },
                "pe": {
                    "type": "number"
                },
                "preclose": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "turnover": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        }
    }
}